refactor: split pressure drop interface
