+++ SPEC.md
@@ -28,7 +28,7 @@
 <<<<<<< HEAD
 | **License**             | MIT                                        |
 | **Current Version**     | N/A                                        |
-| **Spec Version**        | 1.1.14                                     |
+| **Spec Version**        | 1.1.19                                     |
 | **Last Spec Update**    | 2026-04-06                                 |
 
 ## 2. Purpose & Mission
@@ -123,6 +123,7 @@ Tools/
 | UnifiedToolsLauncher | `UnifiedToolsLauncher.py`            | PyQt6 GUI for tool discovery and execution       |
 | Plugin System        | `src/python/plugin_system/`          | Discover, load, and manage plugins               |
 | Shared Utilities     | `src/python/shared_utilities/`       | Common functions, decorators, error handling     |
+| Pressure Drop Calculator | `src/shared/python/upstream_drift_tools/process_calculators/pressure_drop_calculator/` | Facade-driven gas pressure-drop workflows with extracted API, validation, reference, and result-formatting helpers |
 | Engineering Tools    | `src/tools/`                         | 45+ specialized calculation and processing tools |
 | Data Processing      | `src/data_processing/`               | Pipelines, transformers, validators              |
 | Document Processing  | `src/document_processing/`           | PDF extraction, text processing                  |
@@ -265,6 +266,7 @@ Test pyramid with unit tests at the base, integration tests for tool interaction
 - [ ] UnifiedToolsLauncher starts and displays available tools
 - [ ] Plugin discovery finds all registered tools
 - [ ] Calculation produces deterministic results for same inputs
+- [x] Pressure-drop interface regression tests cover facade exports, helper-driven validation, and calculator/model interoperability
 - [ ] Web API endpoint validates input and returns JSON response
 - [ ] Rust kernel outperforms pure Python equivalent by 10x+
 - [ ] Theme system applies consistently across all GUI tools
@@ -475,6 +477,7 @@ Active development with stable core, continuous tool expansion, and web API in p
 | 2026-04-06 | 1.1.16 | Add focus-within styles to video uploader dropzone and missing aria-labels to the volume and seek range inputs in the video processor web application to improve keyboard navigation visibility. |
 | 2026-04-06 | 1.1.17 | Optimize Polynomial Regression Matrix Construction in AnalyticsSuite using single-pass loops. |
 | 2026-04-06 | 1.1.18 | Refactored `applyFilter` inside `useDataProcessor.ts` to pre-allocate buffers and run the mapping in a single loop. |
+| 2026-04-06 | 1.1.19 | Split `pressure_drop_interface.py` into facade-oriented `pressure_drop_api`, `pressure_drop_validation`, `pressure_drop_reference`, and `pressure_drop_results` modules while preserving the public interface and extending regression coverage for the pressure-drop calculator. |
 
 ---
